Move CSS table functions and tests into css_table module
